4 faliures left

+++ lib/sales_engine.rb
@@ -35,7 +35,7 @@ class SalesEngine
     total_items = CSV.read(file_path_item, headers: true, header_converters: :symbol)
     total_items.each do |item|
       i = Item.new({:id => item[:id].to_i, :name => item[:name], :description => item[:description],
-                    :unit_price => BigDecimal.new(item[:unit_price].to_i/100,4), :created_at => item[:created_at],
+                    :unit_price => BigDecimal.new(item[:unit_price].to_f/100,4), :created_at => item[:created_at],
                     :updated_at => item[:updated_at], :merchant_id => item[:merchant_id]})
       it.add_item(i)
     end
